+++ config/api-factory.php
@@ -4,6 +4,7 @@ return [
     'generateRoutes' => true,
     'generateControllers' => true,
     'generateActions' => true,
+    'generatePostmanCollection' => true,
     'generateSwaggerDoc' => true,
     'overrideControllers' => false,
     'overrideActions' => false,
@@ -11,7 +12,7 @@ return [
 
     'outRouteFileName' => 'routes/web.php',
 
-    'uriPrefix' => '/api/v1/',
+    'uriPrefix' => '/api/v1',
 
     'generatedControllerPathPrefix' => 'app/Http/ApiFactory/Controllers',
     'generatedActionPathPrefix' => 'app/Http/ApiFactory/Actions',