+++ src/Classes/ApiFactoryController.php
@@ -3,13 +3,24 @@
 namespace Dminustin\ApiFactory\Classes;
 
 use App\Http\Controllers\Controller;
+use Dminustin\ApiFactory\Exceptions\ClassNotFoundException;
+use Illuminate\Http\Request;
 
 class ApiFactoryController extends Controller
 {
     protected string $actionName;
 
-    protected function getAction(): ApiFactoryAction
+    public function handle(Request $request)
     {
-        return new $this->actionName();
+        $action = $this->getAction($request);
+        return response()->json($action->run($request)->toArray());
+    }
+
+    protected function getAction(Request $request): ApiFactoryAction
+    {
+        if (!class_exists($this->actionName)) {
+            throw new ClassNotFoundException();
+        }
+        return new $this->actionName($request);
     }
 }