+++ src/Classes/GeneratorInfo.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Dminustin\ApiFactory\Classes;
+
+/**
+ * @property string $name
+ * @property string $version
+ */
+class GeneratorInfo extends BaseModel
+{
+    protected string $name = 'Laravel API Factory';
+    protected string $version;
+}