+++ src/Classes/Generators/HttpActionGenerator.php
@@ -8,7 +8,9 @@ use Dminustin\ApiFactory\Classes\RouterConfig\EndPoint;
 
 class HttpActionGenerator extends AbstractGenerator
 {
-    public function generate()
+    protected string $name = 'Action Generator';
+
+    protected function run()
     {
         $template = $this->loadTemplate('api_factory_action');
 
@@ -23,23 +25,6 @@ class HttpActionGenerator extends AbstractGenerator
                 $this->config->generatedActionNSSuffix,
             );
 
-            if (!$this->config->overrideControllers && file_exists($actionAttributes->getFilePath())) {
-                $this->log->warning('Action exists: ' . json_encode([
-                        'Path:' => $route->path,
-                        'Filename:' => $actionAttributes->getFilePath(),
-                        'ClassName:' => $actionAttributes->getClassName(),
-                        'ShortName:' => $actionAttributes->getShortClassName(),
-                    ]));
-                continue;
-            }
-
-            $this->log->info('Generate Action ' . json_encode([
-                    'Path:' => $route->path,
-                    'Filename:' => $actionAttributes->getFilePath(),
-                    'ClassName:' => $actionAttributes->getClassName(),
-                    'ShortName:' => $actionAttributes->getShortClassName(),
-                ]));
-
             $rules = ['['];
             foreach ($route->params as $key => $value) {
                 $rules[] = '        \'' . $key . '\'=>\'' .
@@ -53,10 +38,11 @@ class HttpActionGenerator extends AbstractGenerator
                 'className' => $actionAttributes->getShortClassName(),
                 'actionName' => $actionAttributes->getClassName(),
                 'rules' => implode(PHP_EOL, $rules),
+                'generator' => $this->info->name . ' v.' . $this->info->version
             ];
 
             $rendered = $this->render($template);
-            $this->writeFile($rendered, $actionAttributes->getFilePath());
+            $this->saveFile($actionAttributes->getFilePath(), $rendered, $this->config->overrideActions);
         }
     }
 }