+++ src/Exceptions/ApiFactoryBaseException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Dminustin\ApiFactory\Exceptions;
+
+class ApiFactoryBaseException extends \Exception
+{
+    /*_*/
+}