+++ src/Exceptions/ClassNotFoundException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Dminustin\ApiFactory\Exceptions;
+
+class ClassNotFoundException extends ApiFactoryBaseException
+{
+    /*_*/
+}