+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to `laravel-api-factory` will be documented in this file
 
-## 1.0.0 - 04 Jan, 2023
+## 1.0.0.0 - 02 Jan, 2023
 
 - initial release
+
+## 1.0.0.1 - 04 Jan, 2023
+
+- Added API Response model
+- Fixed some bugs
+- Changed base Actions and Controllers classes
+- Added documentation
+- Added Exceptions
+
+## 1.0.0.1 - 04 Jan, 2023
+- Added Postman collection generator
+- Fixes some bugs
+- Added PHPStan analyzator support